+++ src/main/java/saboroso/saborosoburguer/entities/menuItems/accompaniment/AddOn.java
@@ -1,15 +1,16 @@
-package saboroso.saborosoburguer.entities.menuItems;
+package saboroso.saborosoburguer.entities.menuItems.accompaniment;
 
 import jakarta.persistence.*;
 import lombok.*;
 import saboroso.saborosoburguer.DTO.addOn.AddOnDTO;
+import saboroso.saborosoburguer.entities.menuItems.MenuItem;
 
 @Entity
 @NoArgsConstructor
 @Table
 @Getter
 @Setter
-public class AddOn extends MenuItem{
+public class AddOn extends MenuItem {
 
     public AddOn(AddOnDTO addOnDTO) {
         setTitle(addOnDTO.title());
