+++ src/main/java/saboroso/saborosoburguer/entities/soldItems/accompaniment/AddOnSale.java
@@ -1,0 +1,22 @@
+package saboroso.saborosoburguer.entities.soldItems.accompaniment;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.NoArgsConstructor;
+import saboroso.saborosoburguer.entities.menuItems.accompaniment.AddOn;
+
+@Entity
+@Table
+@NoArgsConstructor
+public class AddOnSale extends BaseAccompanimentSale {
+    @ManyToOne
+    @JoinColumn(name = "addon_id")
+    private AddOn soldAddOn;
+
+    public AddOnSale(AddOn soldAddOn) {
+        this.soldAddOn = soldAddOn;
+        setSoldFor(soldAddOn.getPrice());
+    }
+}
